# backend/packages/api_sucursal/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "parkos-api-sucursal",
    "description": "Branch API \u2014 operador + sync-agent only, offline-first.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/configuracion/configuracion-tolerancias": {
      "get": {
        "tags": [
          "configuracion",
          "configuracion-tolerancias"
        ],
        "summary": "List Endpoint",
        "operationId": "list_endpoint_api_v1_configuracion_configuracion_tolerancias_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-Sucursal-Context",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sucursal-Context"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionToleranciasReadList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "configuracion",
          "configuracion-tolerancias"
        ],
        "summary": "Create Endpoint",
        "operationId": "create_endpoint_api_v1_configuracion_configuracion_tolerancias_post",
        "parameters": [
          {
            "name": "payload",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Payload"
            }
          },
          {
            "name": "X-Sucursal-Context",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sucursal-Context"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionToleranciasRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/configuracion/configuracion-seguridad": {
      "get": {
        "tags": [
          "configuracion",
          "configuracion-seguridad"
        ],
        "summary": "List Endpoint",
        "operationId": "list_endpoint_api_v1_configuracion_configuracion_seguridad_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-Sucursal-Context",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sucursal-Context"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionSeguridadReadList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "configuracion",
          "configuracion-seguridad"
        ],
        "summary": "Create Endpoint",
        "operationId": "create_endpoint_api_v1_configuracion_configuracion_seguridad_post",
        "parameters": [
          {
            "name": "payload",
            "in": "query",
            "required": true,
            "schema": {
              "title": "Payload"
            }
          },
          {
            "name": "X-Sucursal-Context",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sucursal-Context"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionSeguridadRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/configuracion/configuracion-seguridad/efectiva": {
      "get": {
        "tags": [
          "configuracion"
        ],
        "summary": "Effective seguridad config for a branch (per-branch override OR global default)",
        "description": "Resolve the effective ``configuracion_seguridad`` for ``uuid_sucursal``.\n\nPriority: per-branch row (uuid_sucursal = :requested, vigente_hasta IS NULL)\n\u2192 global default (uuid_sucursal IS NULL, vigente_hasta IS NULL)\n\u2192 404 if neither.",
        "operationId": "efectiva_api_v1_configuracion_configuracion_seguridad_efectiva_get",
        "parameters": [
          {
            "name": "uuid_sucursal",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid Sucursal"
            }
          },
          {
            "name": "X-Sucursal-Context",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sucursal-Context"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionSeguridadRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfiguracionSeguridadRead": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "uuid_sucursal": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uuid Sucursal"
          },
          "dias_expiracion_password": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dias Expiracion Password"
          },
          "max_intentos_login": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Intentos Login"
          },
          "minutos_bloqueo_login": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutos Bloqueo Login"
          },
          "vigente_desde": {
            "type": "string",
            "format": "date-time",
            "title": "Vigente Desde"
          },
          "vigente_hasta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigente Hasta"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "sync_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "uuid",
          "uuid_sucursal",
          "dias_expiracion_password",
          "max_intentos_login",
          "minutos_bloqueo_login",
          "vigente_desde",
          "vigente_hasta",
          "estado",
          "created_at",
          "created_by",
          "sync_status"
        ],
        "title": "ConfiguracionSeguridadRead",
        "description": "Read-back for ``prod.configuracion_seguridad``. ``uuid_sucursal``\nNULL marks the global default row."
      },
      "ConfiguracionSeguridadReadList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConfiguracionSeguridadRead"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ConfiguracionSeguridadReadList"
      },
      "ConfiguracionToleranciasRead": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "uuid_sucursal": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uuid Sucursal"
          },
          "tolerancia_efectivo": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tolerancia Efectivo"
          },
          "tolerancia_datafono": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tolerancia Datafono"
          },
          "vigente_desde": {
            "type": "string",
            "format": "date-time",
            "title": "Vigente Desde"
          },
          "vigente_hasta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigente Hasta"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "sync_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "uuid",
          "uuid_sucursal",
          "tolerancia_efectivo",
          "tolerancia_datafono",
          "vigente_desde",
          "vigente_hasta",
          "estado",
          "created_at",
          "created_by",
          "sync_status"
        ],
        "title": "ConfiguracionToleranciasRead",
        "description": "Read-back for ``prod.configuracion_tolerancias``. ``uuid_sucursal``\nNULL marks the global default row."
      },
      "ConfiguracionToleranciasReadList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConfiguracionToleranciasRead"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ConfiguracionToleranciasReadList"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}